+++ main.go
@@ -142,7 +142,7 @@ func getIniParams() (string, error) {
 func getG2config(ctx context.Context) (g2api.G2config, error) {
 	result := g2config.G2config{}
 	moduleName := "Test module name"
-	verboseLogging := 0 // 0 for no Senzing logging; 1 for logging
+	verboseLogging := int64(0) // 0 for no Senzing logging; 1 for logging
 	iniParams, err := getIniParams()
 	if err != nil {
 		return &result, err
@@ -154,7 +154,7 @@ func getG2config(ctx context.Context) (g2api.G2config, error) {
 func getG2configmgr(ctx context.Context) (g2api.G2configmgr, error) {
 	result := g2configmgr.G2configmgr{}
 	moduleName := "Test module name"
-	verboseLogging := 0
+	verboseLogging := int64(0)
 	iniParams, err := getIniParams()
 	if err != nil {
 		return &result, err
@@ -166,7 +166,7 @@ func getG2configmgr(ctx context.Context) (g2api.G2configmgr, error) {
 func getG2diagnostic(ctx context.Context) (g2api.G2diagnostic, error) {
 	result := g2diagnostic.G2diagnostic{}
 	moduleName := "Test module name"
-	verboseLogging := 0
+	verboseLogging := int64(0)
 	iniParams, err := getIniParams()
 	if err != nil {
 		return &result, err
@@ -178,7 +178,7 @@ func getG2diagnostic(ctx context.Context) (g2api.G2diagnostic, error) {
 func getG2engine(ctx context.Context) (g2api.G2engine, error) {
 	result := g2engine.G2engine{}
 	moduleName := "Test module name"
-	verboseLogging := 0
+	verboseLogging := int64(0)
 	iniParams, err := getIniParams()
 	if err != nil {
 		return &result, err
@@ -190,7 +190,7 @@ func getG2engine(ctx context.Context) (g2api.G2engine, error) {
 func getG2product(ctx context.Context) (g2api.G2product, error) {
 	result := g2product.G2product{}
 	moduleName := "Test module name"
-	verboseLogging := 0
+	verboseLogging := int64(0)
 	iniParams, err := getIniParams()
 	if err != nil {
 		return &result, err
